org/pyproject.toml
[project]
name = "org-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "redis (>=7.1.0,<8.0.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "sqlmodel (>=0.0.31,<0.0.32)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "alembic (>=1.18.1,<2.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "python-multipart (>=0.0.21,<0.0.22)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
